+++ crates/test_utils/wasm/wasm_workspace/Cargo.toml
@@ -14,6 +14,7 @@ members = [
     "debug",
     "dna_properties",
     "emit_signal",
+    "post_commit_signal",
     "entry_defs",
     "hash_entry",
     "foo",
