--- crates/test_utils/wasm/wasm_workspace/emit_signal/src/integrity.rs
@@ -1,11 +1,0 @@
-use hdk::prelude::*;
-
-#[hdk_entry_helper]
-pub struct TestEntry(pub String);
-
-#[derive(Serialize, Deserialize)]
-#[hdk_entry_types]
-#[unit_enum(UnitEntryTypes)]
-pub enum EntryTypes {
-    TestEntry(TestEntry),
-}
