+++ crates/test_utils/wasm/wasm_workspace/post_commit_signal/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "test_wasm_post_commit_signal"
+version = "0.0.1"
+authors = [ "thedavidmeister", "thedavidmeister@gmail.com" ]
+edition = "2021"
+
+[lib]
+name = "test_wasm_post_commit_signal"
+crate-type = [ "cdylib", "rlib" ]
+
+[[example]]
+name = "integrity_test_wasm_post_commit_signal"
+path = "src/integrity.rs"
+crate-type = ["cdylib", "rlib"]
+
+
+[dependencies]
+hdk = { path = "../../../../hdk" }
+serde = { workspace = true }
+holochain_test_wasm_common = { path = "../../../wasm_common" }
+
+[features]
+default = []
+mock = ["hdk/mock"]
